requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
pydantic
python-dotenv
passlib[bcrypt]
pyjwt
fastapi-mail
redis
celery
aiofiles
email-validator
